4_boolean_networks/requirements.txt
apptools==5.1.0
argcomplete==1.12.3
astroid==2.4.2
astropy==4.2
astroquery==0.4.1
autopep8==1.5.4
backcall==0.1.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==4.0.0
colorama==0.4.3
colour==0.1.5
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
envisage==5.0.0
future==0.18.2
glcontext==2.1.0
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.1.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
keyring==22.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
manimgl==1.0.0
ManimPango==0.2.5.post0
mapbox-earcut==0.12.7
matplotlib==3.3.4
mayavi==4.7.2
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
moderngl==5.6.0
moderngl-window==2.1.1
mpmath==1.2.1
multipledispatch==0.6.0
numpy==1.19.2
opencv-python==3.4.2.17
parso==0.7.0
pickleshare==0.7.5
Pillow==8.1.0
pipx==0.16.2.1
progressbar==2.5
prompt-toolkit==3.0.5
pycairo==1.19.1
pycodestyle==2.6.0
pydub==0.23.0
pyerfa==1.7.2
pyface==7.2.0
pyglet==1.5.7
Pygments==2.7.1
pylint==2.5.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyreadline==2.1
pyrr==0.10.3
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
scipy==1.3.0
screeninfo==0.6.7
sip==4.19.24
six==1.15.0
soupsieve==2.2
sympy==1.8
toml==0.10.1
tqdm==4.24.0
traitlets==4.3.3
traits==6.2.0
traitsui==7.1.1
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.26.3
userpath==1.5.0
validators==0.18.2
vtk==9.0.1
wcwidth==0.1.9
webencodings==0.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.4.0
